refactor(ndmtmp): migrate forwarder component to v2
