+++ comp/ndmtmp/forwarder/fx/fx.go
@@ -3,22 +3,17 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2023-present Datadog, Inc.
 
-// Package forwarderimpl exposes the event platform forwarder for netflow.
-package forwarderimpl
+// Package fx provides the fx module for the forwarder component.
+package fx
 
 import (
-	"github.com/DataDog/datadog-agent/comp/aggregator/demultiplexer"
-	"github.com/DataDog/datadog-agent/comp/ndmtmp/forwarder"
+	forwarderimpl "github.com/DataDog/datadog-agent/comp/ndmtmp/forwarder/impl"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
-	"go.uber.org/fx"
 )
 
 // Module defines the fx options for this component.
 func Module() fxutil.Module {
 	return fxutil.Component(
-		fx.Provide(getForwarder))
-}
-
-func getForwarder(agg demultiplexer.Component) (forwarder.Component, error) {
-	return agg.GetEventPlatformForwarder()
+		fxutil.ProvideComponentConstructor(forwarderimpl.GetForwarder),
+	)
 }
