+++ comp/netflow/server/integration_test.go
@@ -28,7 +28,7 @@ import (
 
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 
-	"github.com/DataDog/datadog-agent/comp/ndmtmp/forwarder"
+	forwardermock "github.com/DataDog/datadog-agent/comp/ndmtmp/forwarder/mock"
 
 	ndmtestutils "github.com/DataDog/datadog-agent/pkg/networkdevice/testutils"
 
@@ -75,7 +75,7 @@ func assertFlowEventsCount(t *testing.T, port uint16, srv *Server, packetData []
 func TestNetFlow_IntegrationTest_NetFlow5(t *testing.T) {
 	port, err := ndmtestutils.GetFreePort()
 	require.NoError(t, err)
-	var epForwarder forwarder.MockComponent
+	var epForwarder forwardermock.MockComponent
 	srv := fxutil.Test[Component](t, fx.Options(
 		testOptions,
 		fx.Populate(&epForwarder),
@@ -100,7 +100,7 @@ func TestNetFlow_IntegrationTest_NetFlow5(t *testing.T) {
 func TestNetFlow_IntegrationTest_NetFlow9(t *testing.T) {
 	port, err := ndmtestutils.GetFreePort()
 	require.NoError(t, err)
-	var epForwarder forwarder.MockComponent
+	var epForwarder forwardermock.MockComponent
 	srv := fxutil.Test[Component](t, fx.Options(
 		testOptions,
 		fx.Populate(&epForwarder),
@@ -123,7 +123,7 @@ func TestNetFlow_IntegrationTest_NetFlow9(t *testing.T) {
 func TestNetFlow_IntegrationTest_SFlow5(t *testing.T) {
 	port, err := ndmtestutils.GetFreePort()
 	require.NoError(t, err)
-	var epForwarder forwarder.MockComponent
+	var epForwarder forwardermock.MockComponent
 	srv := fxutil.Test[Component](t, fx.Options(
 		testOptions,
 		fx.Populate(&epForwarder),
@@ -146,7 +146,7 @@ func TestNetFlow_IntegrationTest_SFlow5(t *testing.T) {
 func TestNetFlow_IntegrationTest_AdditionalFields(t *testing.T) {
 	port, err := ndmtestutils.GetFreePort()
 	require.NoError(t, err)
-	var epForwarder forwarder.MockComponent
+	var epForwarder forwardermock.MockComponent
 	srv := fxutil.Test[Component](t, fx.Options(
 		testOptions,
 		fx.Populate(&epForwarder),
