+++ comp/netflow/server/server.go
@@ -18,8 +18,8 @@ import (
 	"github.com/DataDog/datadog-agent/comp/core/hostname"
 	log "github.com/DataDog/datadog-agent/comp/core/log/def"
 	"github.com/DataDog/datadog-agent/comp/core/status"
-	"github.com/DataDog/datadog-agent/comp/ndmtmp/forwarder"
-	nfconfig "github.com/DataDog/datadog-agent/comp/netflow/config/def"
+	forwarder "github.com/DataDog/datadog-agent/comp/ndmtmp/forwarder/def"
+	nfconfig "github.com/DataDog/datadog-agent/comp/netflow/config"
 	"github.com/DataDog/datadog-agent/comp/netflow/flowaggregator"
 	rdnsquerier "github.com/DataDog/datadog-agent/comp/rdnsquerier/def"
 	rdnsquerierimplnone "github.com/DataDog/datadog-agent/comp/rdnsquerier/impl-none"
