+++ comp/networkpath/npcollector/npcollectorimpl/npcollector_testutils.go
@@ -31,7 +31,7 @@ import (
 	"github.com/DataDog/datadog-agent/comp/forwarder/defaultforwarder"
 	"github.com/DataDog/datadog-agent/comp/forwarder/eventplatform/eventplatformimpl"
 	"github.com/DataDog/datadog-agent/comp/forwarder/eventplatformreceiver/eventplatformreceiverimpl"
-	"github.com/DataDog/datadog-agent/comp/ndmtmp/forwarder/forwarderimpl"
+	forwardermock "github.com/DataDog/datadog-agent/comp/ndmtmp/forwarder/mock"
 	"github.com/DataDog/datadog-agent/comp/networkpath/npcollector"
 	npmodel "github.com/DataDog/datadog-agent/comp/networkpath/npcollector/model"
 	traceroute "github.com/DataDog/datadog-agent/comp/networkpath/traceroute/def"
@@ -50,7 +50,7 @@ var MockTimeNow = func() time.Time {
 // testOptions is a fx collection of common dependencies for all tests
 var testOptions = fx.Options(
 	Module(),
-	forwarderimpl.MockModule(),
+	forwardermock.MockModule(),
 	demultiplexerimpl.MockModule(),
 	defaultforwarder.MockModule(),
 	eventplatformimpl.MockModule(),
